feat: Add drawer menu and Star Wars About screen
- Installs react-navigation (drawer, native, screens, safe-area-context, gesture-handler, reanimated) and configures them.
- Creates an `assets/text/about.txt` for the about screen content.
- Implements `src/screens/AboutScreen.js` with a Star Wars-style scrolling text animation.
- Refactors main App.js into `src/screens/GameScreen.js`.
- Creates a new root `App.js` to house the DrawerNavigator.
- Adds a menu button (hamburger icon) to `GameScreen.js` to open the drawer.
- The drawer menu includes navigation to the Game screen and the new About screen.
