+++ assets/text/about.txt
@@ -1,0 +1,17 @@
+Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
+
+Phasellus egestas tellus rutrum tellus pellentesque eu. Tincidunt dui ut ornare lectus sit. Pretium vulputate sapien nec sagittis aliquam malesuada bibendum arcu vitae. Nunc consequat interdum varius sit amet mattis. Arcu cursus vitae congue mauris rhoncus aenean vel elit. Neque viverra justo nec ultrices dui sapien eget mi. Eget nunc lobortis mattis aliquam faucibus purus. Quis lectus nulla at volutpat diam ut venenatis.
+
+A long time ago in a galaxy far, far away....
+
+It is a period of civil war. Rebel spaceships, striking from a hidden base, have won their first victory against the evil Galactic Empire.
+
+During the battle, Rebel spies managed to steal secret plans to the Empire's ultimate weapon, the DEATH STAR, an armored space station with enough power to destroy an entire planet.
+
+Pursued by the Empire's sinister agents, Princess Leia races home aboard her starship, custodian of the stolen plans that can save her people and restore freedom to the galaxy....
+
+Porttitor rhoncus dolor purus non enim praesent elementum facilisis. Nibh mauris cursus mattis molestie a iaculis at. Magna fermentum iaculis eu non diam phasellus vestibulum. Eu sem integer vitae justo eget magna fermentum iaculis. Sit amet consectetur adipiscing elit pellentesque habitant morbi tristique senectus. Leo vel fringilla est ullamcorper eget nulla facilisi.
+
+Morbi enim nunc faucibus a pellentesque sit amet porttitor. Ornare arcu dui vivamus arcu felis bibendum ut tristique. Mattis molestie a iaculis at erat pellentesque adipiscing commodo elit. Vitae elementum curabitur vitae nunc sed velit dignissim sodales ut. Tellus elementum sagittis vitae et leo duis ut. Purus sit amet luctus venenatis lectus magna fringilla urna.
+
+Enim neque volutpat ac tincidunt vitae semper quis lectus. Accumsan lacus vel facilisis volutpat est velit egestas. Bibendum est ultricies integer quis auctor elit sed vulputate. Convallis aenean et tortor at risus viverra adipiscing at in. Cras semper auctor neque vitae tempus quam.
